tools/qwenvl/requirements.txt
--extra-index-url https://pypi.nvidia.com
#tensorrt_llm==0.12.0
datasets~=2.16.0
evaluate~=0.4.1
rouge_score~=0.1.2
transformers-stream-generator
sentencepiece~=0.1.99
tiktoken
einops
auto-gptq
matplotlib
#torchvision==0.17.1